Test of particle.enums moved to dedicated test_enums.py file
