+++ tests/particle/test_enums.py
@@ -1,0 +1,16 @@
+# Licensed under a 3-clause BSD style license, see LICENSE.
+
+from __future__ import absolute_import, division, print_function
+
+from particle.particle.enums import Charge, SpinType
+
+
+def test_enums_Charge():
+    assert Charge.p + Charge.m == Charge.o
+    assert Charge.pp + Charge.mm == Charge.o
+
+
+def test_enums_SpinType():
+    assert SpinType.PseudoScalar == - SpinType.Scalar
+    assert SpinType.Axial == - SpinType.Vector
+    assert SpinType.PseudoTensor == - SpinType.Tensor
